requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
insightface>=0.7.3
faiss-cpu>=1.12.0
opencv-python>=4.8.1.78
torch>=2.6.0
torchvision>=0.19.0
numpy>=1.24.3
pillow>=10.1.0
pycryptodome>=3.19.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
onnxruntime>=1.23.0
